Introduce PCFGGrammar and implement Grammar
This introduces a trait called PCFGGrammar that the Grammar struct is
implemented for.
Here functions are implemented that create a normalised PCFG grammar
from an initial and a Vec of parse trees. 
Note that read_rules takes a function that chooses the initial subtree
to read the rules off (This might be useful later, but for now it
returns the root of the a given parse tree).

Also fmt::Display was implemented for Grammar and Rule.

The lsp-generated implementations for PartialEq were replaced by a
Derive macro.